Remove Qt dependencies and fixes
Plugins are not supposed to pull in UI dependencies where possible. This
is one case where creation of a window is all that's done and should be
done using an abstraction layer held within the plugin itself to prevent
the dependency on Qt.

We remove any reliance on Qt structures. A public-domain base64
implementation was added to replace Qt's base64 implementation. We now
use libobs utility functions for searching and we use std::string and
std::vector for container/string manipulation.

In addition, some changes were made to accomadate these changes. In
particular, the button which wasn't functioning previously had to be
fixed since it broke completely for our case for reasons I'm unsure.
What I am sure of is that the button was broken since it didn't store
state of the button and each time the properties were reloaded, the
button would break.

Another issue where we were passing an std::string directly to a blog
function to be replaced in a %s structure caused weird issues was fixed.

Finally, a note that this entire plugin should be rewritten should be
made. It's a mess. Most of my changes feel like a hack and they took
longer than they probably should have and they're platform dependent
meaning these changes will not work for Linux or MacOS right now.