Separate UI from core logic
This separates the UI from the core logic, moving everything related to the VST UI into another thread. There are various cases here where synchronization is needed (create and destroy), but not many actually benefit from it or need it.

Be very careful when modifying anything in the UI message loop, this is very brittle in terms of stability. I also don't know how well VST handles calls to the ->dispatcher across two threads, but I saw no issues with it so far.